Use foreach and cleanup
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@1804125 13f79535-47bb-0310-9956-ffa450edef68

+++ fop-core/src/main/java/org/apache/fop/pdf/PDFPattern.java
@@ -249,7 +249,7 @@ public class PDFPattern extends PDFPathPaint {
                 p.append("/Matrix [ ");
                 for (tempInt = 0; tempInt < vectorSize; tempInt++) {
                     p.append(PDFNumber.doubleOut(
-                            ((Double)this.matrix.get(tempInt)).doubleValue(), 8));
+                            (Double) this.matrix.get(tempInt), 8));
                     p.append(" ");
                 }
                 p.append("] \n");
@@ -304,7 +304,7 @@ public class PDFPattern extends PDFPathPaint {
                 p.append("/Matrix [ ");
                 for (tempInt = 0; tempInt < vectorSize; tempInt++) {
                     p.append(PDFNumber.doubleOut(
-                            ((Double)this.matrix.get(tempInt)).doubleValue(), 8));
+                            (Double) this.matrix.get(tempInt), 8));
                     p.append(" ");
                 }
                 p.append("] \n");
